requirements.txt
pypdf
langchain
langchain_community
langchain-openai
streamlit
streamlit-extras
protobuf~=3.19.0
rank_bm25
tiktoken
faiss-cpu